Owner role (#112)
* add owner role

* owner is worker

* owner repository

* add migrations

* update API

* add dtos

* owner queries

* implement commands

* update seed

* fix missing role

* secure fitness club endpoints

* update fitness club model

* secure fitnessClub commands

* GetFitnessClubOfWorker return also owner club

* add owner to developmnet data

* Abstract mapping

* allow admin to create fitness club

+++ FitnessClubs/FitnessClubs.Application/FitnessClubs/Commands/DeleteFitnessClubCommand.cs
@@ -12,14 +12,31 @@ namespace FitnessClubs.Application.FitnessClubs.Commands
     public class DeleteFitnessClubCommandHandler : IRequestHandler<DeleteFitnessClubCommand, Result<bool>>
     {
         private readonly IFitnessClubRepository _repository;
+        private readonly HttpAuthContext _httpAuthContext;
 
-        public DeleteFitnessClubCommandHandler(IFitnessClubRepository repository)
+        public DeleteFitnessClubCommandHandler(
+            IFitnessClubRepository repository, 
+            HttpAuthContext httpAuthContext)
         {
             _repository = repository;
+            _httpAuthContext = httpAuthContext;
         }
 
         public async Task<Result<bool>> Handle(DeleteFitnessClubCommand request, CancellationToken cancellationToken)
         {
+            var fitnessClub = await _repository.GetById(request.FitnessClubId, false);
+            if (fitnessClub is null)
+            {
+                return new Result<bool>(Common.CommonConsts.NOT_FOUND);
+            }
+
+            // Only fitnessClub owner or admin can delete
+            if (_httpAuthContext.UserId != fitnessClub.OwnerId
+                && _httpAuthContext.UserRole != Common.Enums.RoleType.Administrator)
+            {
+                return new Result<bool>(Common.CommonConsts.Unauthorized);
+            }
+
             var deleteResult = await _repository.Delete(request.FitnessClubId);
 
             if (deleteResult.IsSuccess)
